[config-feature] Added better configuration options
This commit adds the ability to selectively disable different components
of Ultra based on your preferences - for instance, to turn off custom
test output, or nREPL middleware, without affecting the rest of the
plugin.

It also adds error suppression for importing `eastwood.util`, which was
annoying me elsewhere. I don't necessarily think my solution here is the
best one, but it's good enough for the moment.