[release] v0.2.0
This marks the 0.2.0 release. This is a pretty comprehensive refactor,
primarily focused around resolving annoying warnings and bugs from the
java utility functions. The path I've taken is an extremely hacky one,
but as it turns out there is no other acceptable path that works, so I
guess we're a little hosed there.