[test] Bugfix for test initialization
The long and short of this commit is that Leiningen's deliberate
monkeypatching of clojure.test to support `lein retest` is unbelievably
annoying to have to work around.

Consequently, I'm killing the monkeypatching and going for the simplest
route possible after that, which is to provide a direct hook into
clojure.test/run-tests.

I'm not in love with this as a long term solution, but I regard it as
primarily a failure of Leiningen's, that I can potentially work around
in the long run.