[escape-hatch] Custom escape hatch for Datomic types; colorscheme bugfix
This commit adds a new namespace, `ultra.printer.escape-hatch` to allow
me to define custom methods for Puget's `format-doc` multimethod. In
this case I'm using it to add an escape hatch for `datomic.db.Db`
objects; it's a temporary workaround until I can figure out a better
long term solution with @greglook, though it'll work fine for now.

This commit also adds a bugfix for colorscheme selection that was
introduced in `0.3.0`.

This commit is therefore intended to resolve both
https://github.com/venantius/ultra/issues/23 and
https://github.com/venantius/ultra/issues/17