[source] Finally adds syntax-highlighting to `clojure.repl/source`
This commit leverages my latest creation, Glow, which has yet to be
formally released. Suffice it to say that it modifies
`clojure.repl/source` to actually rely on a syntax highlighting engine
rather than Fipp/Whidbey/Puget, which print highlighted Clojure data
structures.

Actually figuring out how to bind the contents of ultra.repl/source to
clojure.repl/source was something of a pain in the neck, and I'm not
totally sure this will end up being the final approach, but it works
well enough for me.

Resolves #15