[ultra] Move clojure dependency to :dev profile
This moves the org.clojure/clojure dependency out of the plugin's main
list of dependencies and into its :dev profile. This is necessary
because Ultra will otherwise attempt to rewrite project maps,
potentially forcing an unwanted Clojure version on the end user.

Unfortunately, Ultra will still engage in what could be considered a
slightly unfriendly behavior here with regard to potential exclusions of
Ultra's dependencies -- I'm feeling a little too lazy to do very much
about that for now, but might want to address the issue in the future.