[ultra] Require nrepl or clojure.tools.nrepl dynamically
clojure.tools.nrepl is now officially deprecated and has been replaced
with nrepl/nrepl. This commit updates all direct references to
clojure.tools.nrepl to use a dynamic import that checks to see if
clojure.tools.nrepl is on the path and otherwise to use the new nrepl.

As we rely on a significant amount of functionality in Whidbey/Puget,
those will also need to have their references updated.