[ultra] Migrate to new nREPL infrastructure
This commit includes some fairly significant changes to Ultra's
underlying nREPL architecture that are a consequence of changes to the
way Lein's repl works as of 2.9.0. Those changes also affect other parts
of this plugin.

The changes in Ultra would not be possible without work done on many
upstream repositories, but @greglook deserves a ton of credit for his
tireless work on Whidbey and Puget.

At the moment there is a remaining niggling issue which is that due to
the move to streaming rendering in the client, an additional newline is
printed after each return. I'm sure some enterprising person will fix
this soon enough.